python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
shioaji[speed]==1.2.9
pyyaml==6.0.1
requests==2.31.0
feedparser==6.0.10
pycryptodome==3.20.0
yfinance>=1.0
urllib3<2.0.0
psycopg2-binary==2.9.9

# Test dependencies
pytest==8.4.1
pytest-cov==7.0.0
